+++ prefect_gcp/bigquery.py
@@ -547,6 +547,7 @@ class BigQueryWarehouse(DatabaseBlock):
 
     _block_type_name = "BigQuery Warehouse"
     _logo_url = "https://images.ctfassets.net/gm98wzqotmnx/4CD4wwbiIKPkZDt4U3TEuW/c112fe85653da054b6d5334ef662bec4/gcp.png?h=250"  # noqa
+    _documentation_url = "https://prefecthq.github.io/prefect-gcp/bigquery/#prefect_gcp.bigquery.BigQueryWarehouse"  # noqa: E501
 
     gcp_credentials: GcpCredentials
     fetch_size: int = Field(