+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- `get_directory` method in `GcsBucket` returns the list of downloaded file paths and supports relative path for `local_path` [#129](https://github.com/PrefectHQ/prefect-gcp/pull/129).
+
 ### Deprecated
 
 ### Removed
@@ -18,6 +20,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 ### Security
+
+## 0.2.5
+
+Released on January 27th, 2023.
+
+### Added
+
+- `vpc_connector_name` field to `CloudRunJob` - [#123](https://github.com/PrefectHQ/prefect-gcp/pull/123)
+- `list_folders` method for GcsBucket - [#121](https://github.com/PrefectHQ/prefect-gcp/pull/121)
+
+### Fixed
+
+- Listing blobs at the root folder - [#120](https://github.com/PrefectHQ/prefect-gcp/pull/120)
 
 ## 0.2.4